feat: add ErrorHandler class for centralized error handling in Avro deserialization
